Implement 64-bit support for the kernel and user space
- Added 64-bit entry point and paging setup in entry64.asm.
- Created a new linker script for the 64-bit kernel (kernel64.ld).
- Implemented context switching for 64-bit processes in swtch64.asm.
- Adapted trap handling for 64-bit architecture in trapasm64.asm.
- Introduced x86_64.asm for x86-specific routines in 64-bit mode.
- Developed initcode64.asm to initialize the user space for 64-bit processes.
- Created usys64.asm for syscall stubs in 64-bit user space.
- Updated printf implementation in printf.c to handle pointer-width values.
- Modified usertests.c to accommodate changes in pointer types.

+++ boot/bootmain.c
@@ -34,21 +34,26 @@ bootmain(void)
   // Load each program segment (ignores ph flags).
   // ph->paddr, not ph->vaddr: the boot loader runs with paging off, so
   // it can only write to physical addresses, and the kernel's linker
-  // script (kernel/kernel.ld) arranges for paddr to be exactly the low
-  // physical load address the kernel's entry code (kernel/entry.asm)
-  // expects to find itself at before it enables paging.
-  ph = (struct proghdr*)((uchar*)elf + elf->phoff);
+  // script (kernel/kernel.ld / kernel/kernel64.ld) arranges for paddr to
+  // be exactly the low physical load address the kernel's entry code
+  // (kernel/entry.asm / kernel/entry64.asm) expects to find itself at
+  // before it enables paging - and, on the 64-bit build, before it's
+  // even reached long mode, so that address is always a plain 32-bit
+  // physical address (< 4GB) despite proghdr's fields being 64-bit wide
+  // there; the (uint) casts below make that truncation explicit rather
+  // than relying on an implicit one.
+  ph = (struct proghdr*)((uchar*)elf + (uint)elf->phoff);
   eph = ph + elf->phnum;
   for(; ph < eph; ph++){
-    pa = (uchar*)ph->paddr;
-    readseg(pa, ph->filesz, ph->off);
+    pa = (uchar*)(uint)ph->paddr;
+    readseg(pa, (uint)ph->filesz, (uint)ph->off);
     if(ph->memsz > ph->filesz)
-      stosb(pa + ph->filesz, 0, ph->memsz - ph->filesz);
+      stosb(pa + (uint)ph->filesz, 0, (uint)(ph->memsz - ph->filesz));
   }
 
   // Call the entry point from the ELF header.
   // Does not return!
-  entry = (void(*)(void))(elf->entry);
+  entry = (void(*)(void))(uint)elf->entry;
   entry();
 }
